requirements.txt
coverage==4.0.3
py==1.4.31
pytest==2.8.7
pytest-cov==2.2.0
PyYAML==3.11
requests==2.9.1
six==1.10.0
vcrpy==1.7.4
wrapt==1.10.6
